backend/functions/manage_tokens/requirements.txt
# === manage_tokens/requirements.txt ===
Flask==2.3.2
flask-cors==4.0.0
firebase-admin==6.3.0
google-cloud-secret-manager==2.16.3
google-cloud-firestore==2.13.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gunicorn